db: call rset::get_precision with wrong type

+++ src/db/dbConvert.c
@@ -1628,7 +1628,7 @@ static long getFloatString(
     char *pbuffer = (char *)pto;
     float	*psrc=(float *)(paddr->pfield);
     long	status = 0;
-    int		precision = 6;
+    long		precision = 6;
     struct rset	*prset = 0;
 
     if(paddr) prset = dbGetRset(paddr);
@@ -1838,7 +1838,7 @@ static long getDoubleString(
     char *pbuffer = (char *)pto;
     double *psrc=(double *)(paddr->pfield);
     long	status = 0;
-    int		precision = 6;
+    long	precision = 6;
     struct rset	*prset = 0;
 
     if(paddr) prset = dbGetRset(paddr);
@@ -3940,7 +3940,7 @@ static long putFloatString(
     const float *pbuffer = (const float *)pfrom;
     char	*pdest=(char *)(paddr->pfield);
     long	status = 0;
-    int		precision = 6;
+    long		precision = 6;
     struct rset *prset = dbGetRset(paddr);
     short size=paddr->field_size;
 
@@ -4150,7 +4150,7 @@ static long putDoubleString(
     const double *pbuffer = (const double *)pfrom;
     char	*pdest=(char *)(paddr->pfield);
     long	status = 0;
-    int		precision = 6;
+    long	precision = 6;
     struct rset *prset = dbGetRset(paddr);
     short size=paddr->field_size;
